Rename package and fix err not returned
