+++ internal/processing/hashbasedstate/desired.go
@@ -1,4 +1,4 @@
-package hashablestate
+package hashbasedstate
 
 import (
 	"fmt"
