+++ internal/processing/hashbasedstate/hashbasedstate.go
@@ -1,17 +1,16 @@
-// Package hashablestate provides types and functions to compare objects by a hash and a position in a yaml sequence.
+// Package hashbasedstate provides types and functions to compare objects by a hash and a position in a yaml sequence.
 //
-// The hashablestate package should only be used objects unmarshalled from a yaml sequences as they have a defined order.
+// The hashbasedstate package should only be used objects unmarshalled from a yaml sequences as they have a defined order.
 // The comparison is based on labels for a hash and an index put on a kubernetes object. The hash label holds the hash that
 // represents the object and the index value holds the position of the object in the sequence. Both of this information is
 // then used to identify if an object was changed, removed or newly added.
 //
 // Since this comparison is based on the order of objects, it means that adding a new object before an existing object in the
 // sequence triggers an update for all the following objects, since their position in the sequence has changed.
-package hashablestate
+package hashbasedstate
 
 import (
 	"fmt"
-	"github.com/kyma-project/api-gateway/internal/helpers"
 	securityv1beta1 "istio.io/client-go/pkg/apis/security/v1beta1"
 	"strconv"
 	"strings"
@@ -35,7 +34,7 @@ func AddHashingLabels(ap *securityv1beta1.AuthorizationPolicy, indexInYaml int) 
 	// is static (https://yaml.org/spec/1.2/spec.html#id2764044).
 	ap.Labels[indexLabelName] = strconv.Itoa(indexInYaml)
 
-	hash, err := helpers.GetAuthorizationPolicyHash(ap)
+	hash, err := GetAuthorizationPolicyHash(ap)
 	if err != nil {
 		return err
 	}
